Move data generator test.
