+++ build.gradle.kts
@@ -16,14 +16,6 @@ dependencies {
     implementation(project(":data"))
     implementation(project(":model"))
     implementation(libs.clikt)
-    implementation(libs.faker)
-    implementation(libs.jooq.jooq)
-    implementation(libs.mariadb)
-    implementation(libs.mysql)
-    implementation(libs.postgres)
-    implementation(libs.testcontainers.mariadb)
-    implementation(libs.testcontainers.mysql)
-    implementation(libs.testcontainers.postgres)
 }
 
 application {
